requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
Deprecated==1.2.13
distro==1.8.0
Django==4.1.5
django-cors-headers==3.13.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djangotoolbox==1.8.0
djoser==2.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
netaddr==0.8.0
oauthlib==3.2.2
Pillow==9.4.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
python3-openid==3.2.0
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
stripe==5.0.0
uritemplate==4.1.1
urllib3==1.26.13
wrapt==1.14.1
yookassa==2.3.4
